+++ device-management-application/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- [SW-15581] Define a default verbosity level for DM_VLOG.
 ### Deprecated
 ### Removed
 ### Fixed